RavenDB-17827 Removed all unnecessary string allocations in IndexSearcher. Cleanup of IndexSearcher.
