+++ src/Corax/Queries/TermProviders/TermProvider.Contains.cs
@@ -14,17 +14,14 @@ namespace Corax.Queries
         private readonly Slice _term;
 
         private CompactTree.Iterator _iterator;
-
-        public ContainsTermProvider(IndexSearcher searcher, ByteStringContext context, CompactTree tree, string field, string term, int fieldId)
+        public ContainsTermProvider(IndexSearcher searcher, ByteStringContext context, CompactTree tree, string field, int fieldId, Slice term)
         {
             _tree = tree;
             _searcher = searcher;
             _field = field;
             _iterator = tree.Iterate();
             _iterator.Reset();
-
-            Slice.From(context, _searcher.EncodeTerm(term, fieldId), out _term);
-           
+            _term = term;
         }
 
         public void Reset()
@@ -41,7 +38,7 @@ namespace Corax.Queries
                 if (!termSlice.Contains(contains))
                     continue;
 
-                term = _searcher.TermQuery(_field, termSlice.ToString());
+                term = _searcher.TermQuery(_field, termSlice);
                 return true;
             }
