+++ src/Corax/Queries/TermProviders/TermProvider.EndsWith.cs
@@ -13,15 +13,14 @@ namespace Corax.Queries
         private readonly Slice _endsWith;
 
         private CompactTree.Iterator _iterator;
-        public EndsWithTermProvider(IndexSearcher searcher, ByteStringContext context, CompactTree tree, string field, int fieldId, string endsWith)
+        public EndsWithTermProvider(IndexSearcher searcher, ByteStringContext context, CompactTree tree, string field, int fieldId, Slice endsWith)
         {
             _tree = tree;
             _searcher = searcher;
             _field = field;
             _iterator = tree.Iterate();
             _iterator.Reset();
-
-            Slice.From(context, _searcher.EncodeTerm(endsWith, fieldId), out _endsWith);
+            _endsWith = endsWith;
         }
 
         public void Reset()
@@ -38,7 +37,7 @@ namespace Corax.Queries
                 if (termSlice.EndsWith(suffix) == false)
                     continue;
 
-                term = _searcher.TermQuery(_field, termSlice.ToString());
+                term = _searcher.TermQuery(_field, termSlice);
                 return true;
             }
