+++ src/Corax/Queries/TermProviders/TermProvider.NotStartsWith.cs
@@ -18,14 +18,13 @@ namespace Corax.Queries
         private readonly string _field;
         private readonly Slice _startWith;
 
-        public NotStartWithTermProvider(IndexSearcher searcher, ByteStringContext context, CompactTree tree, string field, int fieldId, string startWith)
+        public NotStartWithTermProvider(IndexSearcher searcher, ByteStringContext context, CompactTree tree, string field, int fieldId, Slice startWith)
         {
             _searcher = searcher;
             _field = field;
             _iterator = tree.Iterate();
             _iterator.Reset();
-
-            Slice.From(context, _searcher.EncodeTerm(startWith, fieldId), out _startWith);
+            _startWith = startWith;
         }
 
         public void Reset()
@@ -42,7 +41,7 @@ namespace Corax.Queries
                 if (termSlice.StartWith(_startWith))
                     continue;
 
-                term = _searcher.TermQuery(_field, termSlice.ToString());
+                term = _searcher.TermQuery(_field, termSlice);
                 return true;
             }
